Extracted spooky descriptions and calculation of level to a separate module.
